PDF_chat/requirements.txt
aiohttp                 
aiosignal               
aniso8601               
annotated-types         
anyio                   
attrs                    
blinker                 
certifi                  
charset-normalizer      
click                   
contourpy               
cycler                   
dataclasses-json        
distro                  
faiss-cpu               
Flask                   
Flask-RESTful          
Flask-SQLAlchemy        
fonttools                
frozenlist              
greenlet                
h11                      
httpcore                
httpx                    
idna                   
itsdangerous            
Jinja2                  
jsonpatch              
jsonpointer             
kiwisolver              
langchain               
langchain-community     
langchain-core          
langchain-text-splitters
langsmith               
MarkupSafe               
marshmallow           
matplotlib            
multidict            
mypy-extensions         
numpy              
openai             
orjson             
packaging            
pandas           
pillow             
pip                     
pydantic                 
pydantic_core            
pyodbc                  
pyparsing                
PyPDF2                   
python-dateutil          
python-dotenv            
pytz                     
PyYAML                  
requests                
six                     
sniffio                 
SQLAlchemy               
tenacity                 
tqdm                    
typing_extensions        
typing-inspect          
tzdata                   
urllib3                  
Werkzeug                 
yarl                     